refactor: use scss module
