--- apps/home/src/app/components/About.module.css
@@ -1,200 +1,0 @@
-.about {
-  padding: 100px 0;
-  background: #ffffff;
-}
-
-.aboutContent {
-  max-width: 900px;
-  margin: 0 auto;
-}
-
-.aboutText p {
-  font-size: 1.1rem;
-  line-height: 1.8;
-  color: #666;
-  margin-bottom: 1.5rem;
-  text-align: center;
-  word-break: keep-all;
-}
-
-.aboutDetails {
-  margin: 2.5rem 0;
-  display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(280px, 1fr));
-  gap: 1.5rem;
-}
-
-.detailSection {
-  background: linear-gradient(135deg, #f8f9fa 0%, #ffffff 100%);
-  border-radius: 16px;
-  padding: 1.5rem;
-  border: 1px solid #e9ecef;
-  transition: all 0.3s ease;
-  box-shadow: 0 2px 8px rgba(0, 0, 0, 0.04);
-}
-
-.detailSection:hover {
-  transform: translateY(-2px);
-  box-shadow: 0 8px 25px rgba(102, 126, 234, 0.15);
-  border-color: #667eea;
-}
-
-.detailTitle {
-  font-size: 1.1rem;
-  font-weight: 700;
-  color: #333;
-  margin-bottom: 1rem;
-  display: flex;
-  align-items: center;
-  gap: 0.5rem;
-  padding-bottom: 0.5rem;
-  border-bottom: 2px solid #667eea;
-}
-
-.detailItems {
-  display: flex;
-  flex-direction: column;
-  gap: 0.8rem;
-}
-
-.detailItem {
-  display: flex;
-  flex-direction: column;
-  gap: 0.2rem;
-  align-items: flex-start;
-  padding: 0.8rem;
-  background: rgba(102, 126, 234, 0.03);
-  border-radius: 8px;
-  transition: all 0.2s ease;
-}
-
-.detailItem:hover {
-  background: rgba(102, 126, 234, 0.08);
-  transform: translateX(3px);
-}
-
-.detailPeriod {
-  font-size: 0.8rem;
-  color: #667eea;
-  font-weight: 600;
-  min-width: 90px;
-  flex-shrink: 0;
-  background: rgba(102, 126, 234, 0.1);
-  padding: 0.2rem 0.5rem;
-  border-radius: 12px;
-  text-align: center;
-}
-
-.detailContent {
-  display: flex;
-  padding: 0.2rem 0.5rem;
-  flex: 1;
-}
-
-.detailContent strong {
-  color: #333;
-  font-weight: 600;
-  font-size: 0.95rem;
-  display: block;
-  word-break: keep-all;
-}
-
-.detailStatus {
-  color: #667eea;
-  font-size: 0.8rem;
-  font-weight: 500;
-  margin-left: 0.5rem;
-  padding: 0.1rem 0.4rem;
-  border-radius: 8px;
-}
-
-.detailDescription {
-  font-size: 0.8rem;
-  color: #666;
-  margin-top: 0.3rem;
-  font-style: italic;
-  word-break: keep-all;
-  line-height: 1.4;
-}
-
-.aboutHighlights {
-  display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(250px, 1fr));
-  gap: 1.5rem;
-  margin-top: 2.5rem;
-}
-
-.highlightItem {
-  text-align: center;
-  padding: 1.5rem;
-  border-radius: 16px;
-  background: linear-gradient(135deg, #f8f9fa 0%, #ffffff 100%);
-  transition: all 0.3s ease;
-  border: 1px solid #e9ecef;
-  word-break: keep-all;
-  box-shadow: 0 2px 8px rgba(0, 0, 0, 0.04);
-}
-
-.highlightItem:hover {
-  transform: translateY(-3px);
-  box-shadow: 0 10px 30px rgba(102, 126, 234, 0.15);
-  border-color: #667eea;
-}
-
-.highlightItem h3 {
-  font-size: 1.1rem;
-  font-weight: 600;
-  color: #333;
-  margin-bottom: 0.8rem;
-}
-
-.highlightItem p {
-  font-size: 0.9rem;
-  color: #666;
-  line-height: 1.5;
-  margin: 0;
-}
-
-@media (max-width: 768px) {
-  .about {
-    padding: 80px 0;
-  }
-
-  .aboutContent {
-    max-width: 100%;
-    padding: 0 1rem;
-  }
-
-  .aboutDetails {
-    margin: 2rem 0;
-    gap: 1rem;
-    grid-template-columns: 1fr;
-  }
-
-  .detailSection {
-    padding: 1.2rem;
-  }
-
-  .detailItem {
-    flex-direction: column;
-    gap: 0.5rem;
-    align-items: flex-start;
-    padding: 0.6rem;
-  }
-
-  .detailPeriod {
-    min-width: auto;
-    font-size: 0.75rem;
-    align-self: flex-start;
-  }
-
-  .aboutHighlights {
-    grid-template-columns: 1fr;
-    gap: 1rem;
-    margin-top: 2rem;
-  }
-
-  .highlightItem {
-    padding: 1.2rem;
-  }
-}
