+++ apps/home/src/app/components/About.module.scss
@@ -1,0 +1,194 @@
+@use '../styles/mixins' as *;
+
+.about {
+  padding: 100px 0;
+  background: $background-white;
+
+  @include tablet {
+    padding: 80px 0;
+  }
+}
+
+.aboutContent {
+  max-width: 900px;
+  margin: 0 auto;
+
+  @include tablet {
+    max-width: 100%;
+    padding: 0 $spacing-md;
+  }
+}
+
+.aboutText {
+  p {
+    font-size: 1.1rem;
+    line-height: 1.8;
+    color: $text-secondary;
+    margin-bottom: $spacing-lg;
+    text-align: center;
+    word-break: keep-all;
+  }
+}
+
+.aboutDetails {
+  margin: 2.5rem 0;
+  display: grid;
+  grid-template-columns: repeat(auto-fit, minmax(280px, 1fr));
+  gap: $spacing-lg;
+
+  @include tablet {
+    margin: $spacing-xl 0;
+    gap: $spacing-md;
+    grid-template-columns: 1fr;
+  }
+}
+
+.detailSection {
+  @include card-base;
+  background: linear-gradient(135deg, #f8f9fa 0%, $background-white 100%);
+  padding: $spacing-lg;
+  border: 1px solid #e9ecef;
+  @include card-hover;
+
+  &:hover {
+    border-color: $primary-purple;
+  }
+
+  @include tablet {
+    padding: 1.2rem;
+  }
+}
+
+.detailTitle {
+  font-size: 1.1rem;
+  font-weight: 700;
+  color: $text-primary;
+  margin-bottom: $spacing-md;
+  @include flex-center;
+  justify-content: flex-start;
+  gap: $spacing-sm;
+  padding-bottom: $spacing-sm;
+  border-bottom: 2px solid $primary-purple;
+}
+
+.detailItems {
+  @include flex-column;
+  gap: 0.8rem;
+}
+
+.detailItem {
+  @include flex-column;
+  gap: 0.2rem;
+  align-items: flex-start;
+  padding: 0.8rem;
+  background: $primary-purple-light;
+  border-radius: $border-radius-sm;
+  transition: $transition-fast;
+
+  &:hover {
+    background: rgba(102, 126, 234, 0.08);
+    transform: translateX(3px);
+  }
+
+  @include tablet {
+    flex-direction: column;
+    gap: $spacing-sm;
+    align-items: flex-start;
+    padding: 0.6rem;
+  }
+}
+
+.detailPeriod {
+  font-size: 0.8rem;
+  color: $primary-purple;
+  font-weight: 600;
+  min-width: 90px;
+  flex-shrink: 0;
+  background: rgba(102, 126, 234, 0.1);
+  padding: 0.2rem $spacing-sm;
+  border-radius: $border-radius-md;
+  text-align: center;
+
+  @include tablet {
+    min-width: auto;
+    font-size: 0.75rem;
+    align-self: flex-start;
+  }
+}
+
+.detailContent {
+  display: flex;
+  padding: 0.2rem $spacing-sm;
+  flex: 1;
+
+  strong {
+    color: $text-primary;
+    font-weight: 600;
+    font-size: 0.95rem;
+    display: block;
+    word-break: keep-all;
+  }
+}
+
+.detailStatus {
+  color: $primary-purple;
+  font-size: 0.8rem;
+  font-weight: 500;
+  margin-left: $spacing-sm;
+  padding: 0.1rem 0.4rem;
+  border-radius: $border-radius-sm;
+}
+
+.detailDescription {
+  font-size: 0.8rem;
+  color: $text-secondary;
+  margin-top: 0.3rem;
+  font-style: italic;
+  word-break: keep-all;
+  line-height: 1.4;
+}
+
+.aboutHighlights {
+  display: grid;
+  grid-template-columns: repeat(auto-fit, minmax(250px, 1fr));
+  gap: $spacing-lg;
+  margin-top: 2.5rem;
+
+  @include tablet {
+    grid-template-columns: 1fr;
+    gap: $spacing-md;
+    margin-top: $spacing-xl;
+  }
+}
+
+.highlightItem {
+  @include card-base;
+  text-align: center;
+  padding: $spacing-lg;
+  background: linear-gradient(135deg, #f8f9fa 0%, $background-white 100%);
+  border: 1px solid #e9ecef;
+  word-break: keep-all;
+  @include hover-lift;
+
+  &:hover {
+    border-color: $primary-purple;
+  }
+
+  h3 {
+    font-size: 1.1rem;
+    font-weight: 600;
+    color: $text-primary;
+    margin-bottom: 0.8rem;
+  }
+
+  p {
+    font-size: 0.9rem;
+    color: $text-secondary;
+    line-height: 1.5;
+    margin: 0;
+  }
+
+  @include tablet {
+    padding: 1.2rem;
+  }
+}
