+++ apps/home/src/app/components/About.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import styles from './About.module.css';
+import styles from './About.module.scss';
 
 const About: React.FC = () => {
   return (
