--- apps/home/src/app/components/Contact.module.css
@@ -1,243 +1,0 @@
-.contact {
-  padding: 100px 0;
-  background: linear-gradient(135deg, #667eea 0%, #764ba2 100%);
-  color: white;
-  position: relative;
-  overflow: hidden;
-}
-
-.contact::before {
-  content: '';
-  position: absolute;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  background: url('data:image/svg+xml,<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 100 100"><defs><pattern id="grain" width="100" height="100" patternUnits="userSpaceOnUse"><circle cx="25" cy="25" r="1" fill="rgba(255,255,255,0.1)"/><circle cx="75" cy="75" r="1" fill="rgba(255,255,255,0.1)"/><circle cx="50" cy="10" r="0.5" fill="rgba(255,255,255,0.05)"/><circle cx="10" cy="60" r="0.5" fill="rgba(255,255,255,0.05)"/><circle cx="90" cy="40" r="0.5" fill="rgba(255,255,255,0.05)"/></pattern></defs><rect width="100" height="100" fill="url(%23grain)"/></svg>');
-  pointer-events: none;
-}
-
-.sectionTitle {
-  color: white !important;
-  text-align: center;
-  margin-bottom: 3rem;
-  position: relative;
-  z-index: 1;
-}
-
-.sectionTitle::after {
-  background: rgba(255, 255, 255, 0.3);
-}
-
-.contactContent {
-  position: relative;
-  z-index: 1;
-}
-
-.contactIntro {
-  text-align: center;
-  margin-bottom: 4rem;
-  max-width: 600px;
-  margin-left: auto;
-  margin-right: auto;
-}
-
-.contactIntro h3 {
-  font-size: 2rem;
-  font-weight: 700;
-  margin-bottom: 1rem;
-  color: white;
-  background: linear-gradient(45deg, #fff, #e0e7ff);
-  background-clip: text;
-  -webkit-background-clip: text;
-  -webkit-text-fill-color: transparent;
-}
-
-.contactIntro p {
-  font-size: 1.1rem;
-  color: rgba(255, 255, 255, 0.9);
-  line-height: 1.6;
-}
-
-.contactCards {
-  display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(300px, 1fr));
-  gap: 2rem;
-  margin-bottom: 3rem;
-}
-
-.contactCard {
-  background: rgba(255, 255, 255, 0.1);
-  border: 1px solid rgba(255, 255, 255, 0.2);
-  border-radius: 20px;
-  padding: 2.5rem 2rem;
-  text-align: center;
-  backdrop-filter: blur(20px);
-  transition: all 0.3s ease;
-  position: relative;
-  overflow: hidden;
-}
-
-.contactCard::before {
-  content: '';
-  position: absolute;
-  top: 0;
-  left: -100%;
-  width: 100%;
-  height: 100%;
-  background: linear-gradient(
-    90deg,
-    transparent,
-    rgba(255, 255, 255, 0.1),
-    transparent
-  );
-  transition: left 0.5s ease;
-}
-
-.contactCard:hover::before {
-  left: 100%;
-}
-
-.contactCard:hover {
-  transform: translateY(-10px);
-  background: rgba(255, 255, 255, 0.15);
-  border-color: rgba(255, 255, 255, 0.3);
-  box-shadow: 0 20px 40px rgba(0, 0, 0, 0.1);
-}
-
-.contactIcon {
-  font-size: 3rem;
-  width: 80px;
-  height: 80px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  background: rgba(255, 255, 255, 0.2);
-  border-radius: 50%;
-  margin: 0 auto 1.5rem;
-  transition: all 0.3s ease;
-  color: white;
-}
-
-.contactIcon svg {
-  width: 32px;
-  height: 32px;
-  stroke: currentColor;
-  fill: none;
-}
-
-.contactCard:hover .contactIcon {
-  transform: scale(1.1);
-  background: rgba(255, 255, 255, 0.3);
-}
-
-.contactCard h4 {
-  font-size: 1.3rem;
-  font-weight: 700;
-  margin-bottom: 0.5rem;
-  color: white;
-}
-
-.contactCard p {
-  color: rgba(255, 255, 255, 0.8);
-  font-size: 1rem;
-  margin-bottom: 1.5rem;
-}
-
-.contactAction {
-  display: inline-block;
-  padding: 0.75rem 1.5rem;
-  background: rgba(255, 255, 255, 0.2);
-  color: white;
-  text-decoration: none;
-  border-radius: 50px;
-  font-weight: 600;
-  transition: all 0.3s ease;
-  border: 2px solid rgba(255, 255, 255, 0.3);
-}
-
-.contactAction:hover {
-  background: rgba(255, 255, 255, 0.3);
-  border-color: rgba(255, 255, 255, 0.5);
-  transform: translateY(-2px);
-}
-
-.professionalLinks {
-  display: flex;
-  gap: 0.5rem;
-  justify-content: center;
-}
-
-.proLink {
-  padding: 0.5rem 1rem;
-  background: rgba(255, 255, 255, 0.1);
-  color: white;
-  text-decoration: none;
-  border-radius: 20px;
-  font-size: 0.9rem;
-  font-weight: 500;
-  transition: all 0.3s ease;
-  border: 1px solid rgba(255, 255, 255, 0.2);
-  display: flex;
-  align-items: center;
-  gap: 0.5rem;
-}
-
-.proLink:hover {
-  background: rgba(255, 255, 255, 0.2);
-  transform: translateY(-2px);
-}
-
-.githubLink:hover {
-  background: rgba(33, 41, 60, 0.8);
-  border-color: rgba(33, 41, 60, 0.5);
-}
-
-.linkedinLink:hover {
-  background: rgba(10, 102, 194, 0.8);
-  border-color: rgba(10, 102, 194, 0.5);
-}
-
-.proLink svg {
-  width: 16px;
-  height: 16px;
-  fill: currentColor;
-}
-
-@media (max-width: 768px) {
-  .contact {
-    padding: 80px 0;
-  }
-
-  .contactIntro h3 {
-    font-size: 1.5rem;
-  }
-
-  .contactIntro p {
-    font-size: 1rem;
-  }
-
-  .contactCards {
-    grid-template-columns: 1fr;
-    gap: 1.5rem;
-  }
-
-  .contactCard {
-    padding: 2rem 1.5rem;
-  }
-
-  .contactIcon {
-    font-size: 2.5rem;
-    width: 60px;
-    height: 60px;
-  }
-
-  .professionalLinks {
-    flex-direction: column;
-    gap: 0.75rem;
-  }
-
-  .proLink {
-    padding: 0.75rem 1rem;
-  }
-}
