+++ apps/home/src/app/components/Contact.module.scss
@@ -1,0 +1,222 @@
+@use '../styles/mixins' as *;
+
+.contact {
+  padding: 100px 0;
+  background: linear-gradient(135deg, $primary-purple 0%, #764ba2 100%);
+  color: white;
+  position: relative;
+  overflow: hidden;
+
+  &::before {
+    content: '';
+    position: absolute;
+    top: 0;
+    left: 0;
+    right: 0;
+    bottom: 0;
+    background: url('data:image/svg+xml,<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 100 100"><defs><pattern id="grain" width="100" height="100" patternUnits="userSpaceOnUse"><circle cx="25" cy="25" r="1" fill="rgba(255,255,255,0.1)"/><circle cx="75" cy="75" r="1" fill="rgba(255,255,255,0.1)"/><circle cx="50" cy="10" r="0.5" fill="rgba(255,255,255,0.05)"/><circle cx="10" cy="60" r="0.5" fill="rgba(255,255,255,0.05)"/><circle cx="90" cy="40" r="0.5" fill="rgba(255,255,255,0.05)"/></pattern></defs><rect width="100" height="100" fill="url(%23grain)"/></svg>');
+    pointer-events: none;
+  }
+
+  @include tablet {
+    padding: 80px 0;
+  }
+}
+
+.sectionTitle {
+  color: white !important;
+  text-align: center;
+  margin-bottom: $spacing-xxl;
+  position: relative;
+  z-index: $z-index-base;
+
+  &::after {
+    background: rgba(255, 255, 255, 0.3) !important;
+  }
+}
+
+.contactContent {
+  position: relative;
+  z-index: $z-index-base;
+}
+
+.contactIntro {
+  text-align: center;
+  margin-bottom: 4rem;
+  max-width: 600px;
+  margin-left: auto;
+  margin-right: auto;
+
+  h3 {
+    font-size: 2rem;
+    font-weight: 700;
+    margin-bottom: $spacing-md;
+    color: white;
+    background: linear-gradient(45deg, #fff, #e0e7ff);
+    background-clip: text;
+    -webkit-background-clip: text;
+    -webkit-text-fill-color: transparent;
+
+    @include tablet {
+      font-size: 1.6rem;
+    }
+  }
+
+  p {
+    font-size: 1.1rem;
+    color: rgba(255, 255, 255, 0.9);
+    line-height: 1.6;
+
+    @include tablet {
+      font-size: 1rem;
+    }
+  }
+}
+
+.contactCards {
+  display: grid;
+  grid-template-columns: repeat(auto-fit, minmax(300px, 1fr));
+  gap: $spacing-xl;
+  margin-bottom: $spacing-xxl;
+
+  @include tablet {
+    grid-template-columns: 1fr;
+    gap: $spacing-lg;
+  }
+}
+
+.contactCard {
+  @include glassmorphism(0.1);
+  border-radius: $border-radius-xl;
+  padding: 2.5rem $spacing-xl;
+  text-align: center;
+  transition: $transition-normal;
+  position: relative;
+  overflow: hidden;
+
+  &::before {
+    content: '';
+    position: absolute;
+    top: 0;
+    left: -100%;
+    width: 100%;
+    height: 100%;
+    background: linear-gradient(
+      90deg,
+      transparent,
+      rgba(255, 255, 255, 0.1),
+      transparent
+    );
+    transition: left $transition-slow;
+  }
+
+  h4 {
+    font-size: 1.3rem;
+    font-weight: 700;
+    margin-bottom: $spacing-sm;
+    color: white;
+  }
+
+  p {
+    color: rgba(255, 255, 255, 0.8);
+    font-size: 1rem;
+    margin-bottom: $spacing-lg;
+  }
+
+  @include tablet {
+    padding: $spacing-xl $spacing-lg;
+  }
+}
+
+.contactIcon {
+  font-size: 3rem;
+  width: 80px;
+  height: 80px;
+  @include flex-center;
+  background: rgba(255, 255, 255, 0.2);
+  border-radius: $border-radius-full;
+  margin: 0 auto $spacing-lg;
+  transition: $transition-normal;
+  color: white;
+
+  svg {
+    width: 32px;
+    height: 32px;
+    stroke: currentColor;
+    fill: none;
+  }
+
+  @include tablet {
+    width: 60px;
+    height: 60px;
+    font-size: 2.5rem;
+
+    svg {
+      width: 24px;
+      height: 24px;
+    }
+  }
+}
+
+.contactAction {
+  display: inline-block;
+  padding: 0.75rem $spacing-lg;
+  background: rgba(255, 255, 255, 0.2);
+  color: white;
+  text-decoration: none;
+  border-radius: 50px;
+  font-weight: 600;
+  transition: $transition-normal;
+  border: 2px solid rgba(255, 255, 255, 0.3);
+
+  &:hover {
+    background: rgba(255, 255, 255, 0.3);
+    border-color: rgba(255, 255, 255, 0.5);
+    transform: translateY(-2px);
+  }
+}
+
+.professionalLinks {
+  display: flex;
+  gap: $spacing-sm;
+  justify-content: center;
+
+  @include tablet {
+    flex-direction: column;
+    gap: $spacing-sm;
+  }
+}
+
+.proLink {
+  padding: $spacing-sm $spacing-md;
+  background: rgba(255, 255, 255, 0.1);
+  color: white;
+  text-decoration: none;
+  border-radius: $border-radius-xl;
+  font-size: 0.9rem;
+  font-weight: 500;
+  transition: $transition-normal;
+  border: 1px solid rgba(255, 255, 255, 0.2);
+  @include flex-center;
+  gap: $spacing-sm;
+
+  &:hover {
+    background: rgba(255, 255, 255, 0.2);
+    transform: translateY(-2px);
+  }
+
+  &.githubLink:hover {
+    background: rgba(33, 41, 60, 0.8);
+    border-color: rgba(33, 41, 60, 0.5);
+  }
+
+  &.linkedinLink:hover {
+    background: rgba(10, 102, 194, 0.8);
+    border-color: rgba(10, 102, 194, 0.5);
+  }
+
+  @include tablet {
+    padding: 0.6rem $spacing-md;
+    font-size: 0.85rem;
+  }
+}
