+++ apps/home/src/app/components/Contact.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import styles from './Contact.module.css';
+import styles from './Contact.module.scss';
 
 const Contact: React.FC = () => {
   return (
