+++ apps/home/src/app/components/Footer.module.scss
@@ -1,0 +1,53 @@
+@use '../styles/mixins' as *;
+
+.footer {
+  background: linear-gradient(135deg, #1a1a2e 0%, #16213e 50%, #0f3460 100%);
+  color: white;
+  padding: 30px 0;
+  margin-top: auto;
+  position: relative;
+  overflow: hidden;
+
+  &::before {
+    content: '';
+    position: absolute;
+    top: 0;
+    left: 0;
+    right: 0;
+    bottom: 0;
+    background: url('data:image/svg+xml,<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 100 100"><defs><pattern id="footerGrain" width="100" height="100" patternUnits="userSpaceOnUse"><circle cx="20" cy="20" r="0.5" fill="rgba(255,255,255,0.03)"/><circle cx="80" cy="80" r="0.5" fill="rgba(255,255,255,0.03)"/><circle cx="50" cy="10" r="0.3" fill="rgba(255,255,255,0.02)"/><circle cx="10" cy="70" r="0.3" fill="rgba(255,255,255,0.02)"/><circle cx="90" cy="30" r="0.3" fill="rgba(255,255,255,0.02)"/></pattern></defs><rect width="100" height="100" fill="url(%23footerGrain)"/></svg>');
+    pointer-events: none;
+  }
+
+  @include tablet {
+    padding: 25px 0;
+  }
+
+  @include mobile {
+    padding: 20px 0;
+  }
+}
+
+.footerContent {
+  position: relative;
+  z-index: $z-index-base;
+}
+
+.footerInfo {
+  text-align: center;
+
+  p {
+    color: rgba(255, 255, 255, 0.7);
+    font-size: 0.9rem;
+    margin: 0;
+    font-weight: 400;
+
+    @include tablet {
+      font-size: 0.8rem;
+    }
+
+    @include mobile {
+      font-size: 0.75rem;
+    }
+  }
+}
