+++ apps/home/src/app/components/Footer.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import styles from './Footer.module.css';
+import styles from './Footer.module.scss';
 
 const Footer: React.FC = () => {
   const currentYear = new Date().getFullYear();
