--- apps/home/src/app/components/ImageSlider.module.css
@@ -1,215 +1,0 @@
-.imageSlider {
-  width: 100%;
-  margin-bottom: 3rem;
-}
-
-.sliderContainer {
-  position: relative;
-  width: 100%;
-  max-width: 900px;
-  margin: 0 auto;
-  border-radius: 16px;
-  overflow: hidden;
-  box-shadow: 6px 6px 12px rgba(0, 0, 0, 0.3);
-}
-
-.sliderContent {
-  width: 100%;
-  aspect-ratio: 16 / 9;
-  overflow: hidden;
-  position: relative;
-}
-
-.sliderImage {
-  width: 100%;
-  height: 100%;
-  object-fit: contain;
-  display: block;
-}
-
-.sliderButton {
-  position: absolute;
-  top: 50%;
-  transform: translateY(-50%);
-  background: rgba(0, 0, 0, 0.8);
-  color: white;
-  border: none;
-  width: 50px;
-  height: 50px;
-  border-radius: 50%;
-  font-size: 24px;
-  cursor: pointer;
-  z-index: 2;
-  transition: all 0.3s ease;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  backdrop-filter: blur(10px);
-  opacity: 0;
-  pointer-events: none;
-}
-
-.imageSlider:hover .sliderButton {
-  opacity: 1;
-  pointer-events: auto;
-}
-
-.sliderButton:hover {
-  background: rgba(0, 0, 0, 0.9);
-  transform: translateY(-50%) scale(1.1);
-}
-
-.prev {
-  left: 20px;
-}
-
-.next {
-  right: 20px;
-}
-
-.sliderDots {
-  position: absolute;
-  bottom: 1rem;
-  left: 50%;
-  transform: translateX(-50%);
-  display: flex;
-  justify-content: center;
-  gap: 8px;
-  background: rgba(0, 0, 0, 0.8);
-  padding: 8px 16px;
-  border-radius: 20px;
-  backdrop-filter: blur(10px);
-  opacity: 0;
-  transition: opacity 0.3s ease;
-}
-
-.imageSlider:hover .sliderDots {
-  opacity: 1;
-}
-
-.dot {
-  width: 12px;
-  height: 12px;
-  border-radius: 50%;
-  border: none;
-  background: rgba(255, 255, 255, 0.5);
-  cursor: pointer;
-  transition: all 0.3s ease;
-}
-
-.dot.active {
-  background: white;
-  transform: scale(1.1);
-}
-
-.dot:hover {
-  background: rgba(255, 255, 255, 0.75);
-}
-
-.sliderCounter {
-  text-align: center;
-  margin-top: 1rem;
-  color: #666;
-  font-size: 0.9rem;
-  font-weight: 500;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  gap: 0.5rem;
-  min-height: 2rem;
-}
-
-.autoSlideIndicator {
-  color: #667eea;
-  font-size: 0.8rem;
-  background: rgba(102, 126, 234, 0.1);
-  padding: 0.25rem 0.5rem;
-  border-radius: 12px;
-  border: 1px solid rgba(102, 126, 234, 0.2);
-  transition: opacity 0.3s ease;
-  white-space: nowrap;
-}
-
-.autoSlideIndicator.visible {
-  opacity: 1;
-  animation: pulse 2s infinite;
-}
-
-.autoSlideIndicator.hidden {
-  opacity: 0;
-  animation: none;
-}
-
-@keyframes pulse {
-  0%,
-  100% {
-    opacity: 1;
-  }
-  50% {
-    opacity: 0.7;
-  }
-}
-
-@media (max-width: 768px) {
-  .sliderContainer {
-    max-width: 100%;
-    margin: 0 1rem;
-  }
-
-  .sliderButton {
-    width: 40px;
-    height: 40px;
-    font-size: 20px;
-  }
-
-  .prev {
-    left: 15px;
-  }
-
-  .next {
-    right: 15px;
-  }
-
-  .dot {
-    width: 10px;
-    height: 10px;
-  }
-
-  .sliderDots {
-    padding: 6px 12px;
-  }
-
-  .sliderCounter {
-    flex-direction: column;
-    gap: 0.25rem;
-    min-height: 3rem;
-  }
-
-  .autoSlideIndicator {
-    font-size: 0.75rem;
-  }
-}
-
-@media (max-width: 480px) {
-  .sliderDots {
-    display: none;
-  }
-
-  .sliderContainer {
-    margin: 0 0.5rem;
-  }
-
-  .sliderButton {
-    width: 35px;
-    height: 35px;
-    font-size: 18px;
-  }
-
-  .prev {
-    left: 10px;
-  }
-
-  .next {
-    right: 10px;
-  }
-}
