+++ apps/home/src/app/components/ProjectLayout.module.scss
@@ -1,3 +1,5 @@
+@use '../styles/mixins' as *;
+
 .projectLayout {
   min-height: 100vh;
   background: linear-gradient(135deg, #f5f7fa 0%, #c3cfe2 100%);
