--- apps/home/src/app/components/Projects.module.css
@@ -1,179 +1,0 @@
-.projects {
-  padding: 100px 0;
-  background: linear-gradient(135deg, #f5f7fa 0%, #c3cfe2 100%);
-}
-
-.sectionTitle {
-  text-align: center;
-  font-size: 2.5rem;
-  font-weight: 800;
-  color: #2d3748;
-  margin-bottom: 3rem;
-  position: relative;
-}
-
-.sectionTitle::after {
-  content: '';
-  position: absolute;
-  bottom: -10px;
-  left: 50%;
-  transform: translateX(-50%);
-  width: 60px;
-  height: 4px;
-  background: linear-gradient(135deg, #667eea 0%, #764ba2 100%);
-  border-radius: 2px;
-}
-
-.projectsGrid {
-  display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(350px, 1fr));
-  gap: 2rem;
-}
-
-.projectCard {
-  display: flex;
-  flex-direction: column;
-  background: #ffffff;
-  border-radius: 12px;
-  overflow: hidden;
-  box-shadow: 0 5px 20px rgba(0, 0, 0, 0.1);
-  transition: all 0.3s ease;
-  border: 1px solid #e9ecef;
-}
-
-.projectCard:hover {
-  transform: translateY(-10px);
-  box-shadow: 0 20px 40px rgba(0, 0, 0, 0.15);
-}
-
-.projectImage {
-  height: 200px;
-  background: linear-gradient(135deg, #f5f7fa 0%, #c3cfe2 100%);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  position: relative;
-  overflow: hidden;
-}
-
-.projectImage img {
-  width: 100%;
-  height: 100%;
-  object-fit: cover;
-  object-position: center;
-  transition: transform 0.3s ease;
-}
-
-.projectImage:hover img {
-  transform: scale(1.05);
-}
-
-.imagePlaceholder {
-  color: #666;
-  font-size: 1rem;
-  font-weight: 500;
-}
-
-.projectContent {
-  position: relative;
-  flex: 1;
-  padding: 1.5rem;
-}
-
-.projectTitle {
-  font-size: 1.25rem;
-  font-weight: 700;
-  color: #333;
-  margin-bottom: 0.75rem;
-}
-
-.projectDescription {
-  min-height: 8rem;
-  color: #666;
-  line-height: 1.6;
-  margin-bottom: 1rem;
-  font-size: 0.95rem;
-  word-break: keep-all;
-}
-
-.projectTechnologies {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 0.5rem;
-  margin-bottom: 3.5rem;
-}
-
-.techTag {
-  background: #f8f9fa;
-  color: #667eea;
-  padding: 0.25rem 0.75rem;
-  border-radius: 20px;
-  font-size: 0.8rem;
-  font-weight: 500;
-  border: 1px solid #e9ecef;
-}
-
-.projectLinks {
-  display: flex;
-  position: absolute;
-  bottom: 1.5rem;
-  left: 1.5rem;
-  gap: 0.5rem;
-}
-
-.projectLink {
-  padding: 0.5rem 1rem;
-  height: 100%;
-  border-radius: 6px;
-  text-decoration: none;
-  font-size: 0.9rem;
-  font-weight: 600;
-  transition: all 0.3s ease;
-  border: 2px solid #e9ecef;
-  color: #666;
-}
-
-.projectLink:hover {
-  transform: translateY(-2px);
-  border-color: #667eea;
-  color: #667eea;
-}
-
-.primary {
-  background: linear-gradient(135deg, #667eea 0%, #764ba2 100%);
-  color: white;
-  border-color: transparent;
-}
-
-.primary:hover {
-  box-shadow: 0 5px 15px rgba(102, 126, 234, 0.3);
-  color: white;
-}
-
-@media (max-width: 768px) {
-  .projects {
-    padding: 80px 0;
-  }
-
-  .projectsGrid {
-    grid-template-columns: 1fr;
-    gap: 1.5rem;
-  }
-
-  .projectImage {
-    height: 180px;
-  }
-
-  .projectContent {
-    padding: 1.25rem;
-  }
-
-  .projectLinks {
-    flex-direction: column;
-    gap: 0.75rem;
-  }
-
-  .projectLink {
-    text-align: center;
-  }
-}
