+++ apps/home/src/app/components/Projects.module.scss
@@ -1,0 +1,174 @@
+@use '../styles/mixins' as *;
+
+.projects {
+  padding: 100px 0;
+  background: linear-gradient(135deg, #f5f7fa 0%, #c3cfe2 100%);
+
+  @include tablet {
+    padding: 80px 0;
+  }
+}
+
+.sectionTitle {
+  text-align: center;
+  font-size: 2.5rem;
+  font-weight: 800;
+  color: #2d3748;
+  margin-bottom: $spacing-xxl;
+  position: relative;
+
+  &::after {
+    content: '';
+    position: absolute;
+    bottom: -10px;
+    left: 50%;
+    transform: translateX(-50%);
+    width: 60px;
+    height: 4px;
+    background: linear-gradient(135deg, $primary-purple 0%, #764ba2 100%);
+    border-radius: 2px;
+  }
+}
+
+.projectsGrid {
+  display: grid;
+  grid-template-columns: repeat(auto-fit, minmax(350px, 1fr));
+  gap: $spacing-xl;
+
+  @include tablet {
+    grid-template-columns: 1fr;
+    gap: $spacing-lg;
+  }
+}
+
+.projectCard {
+  @include flex-column;
+  @include card-base;
+  overflow: hidden;
+  border: 1px solid #e9ecef;
+  @include hover-lift;
+
+  &:hover {
+    transform: translateY(-10px);
+    box-shadow: 0 20px 40px rgba(0, 0, 0, 0.15);
+  }
+}
+
+.projectImage {
+  height: 200px;
+  background: linear-gradient(135deg, #f5f7fa 0%, #c3cfe2 100%);
+  @include flex-center;
+  position: relative;
+  overflow: hidden;
+
+  img {
+    width: 100%;
+    height: 100%;
+    object-fit: cover;
+    object-position: center;
+    transition: transform $transition-normal;
+  }
+
+  &:hover img {
+    transform: scale(1.05);
+  }
+
+  @include tablet {
+    height: 180px;
+  }
+}
+
+.imagePlaceholder {
+  color: $text-secondary;
+  font-size: 1rem;
+  font-weight: 500;
+}
+
+.projectContent {
+  position: relative;
+  flex: 1;
+  padding: $spacing-lg;
+
+  @include tablet {
+    padding: 1.25rem;
+  }
+}
+
+.projectTitle {
+  font-size: 1.25rem;
+  font-weight: 700;
+  color: $text-primary;
+  margin-bottom: 0.75rem;
+}
+
+.projectDescription {
+  min-height: 8rem;
+  color: $text-secondary;
+  line-height: 1.6;
+  margin-bottom: $spacing-md;
+  font-size: 0.95rem;
+  word-break: keep-all;
+}
+
+.projectTechnologies {
+  display: flex;
+  flex-wrap: wrap;
+  gap: $spacing-sm;
+  margin-bottom: 3.5rem;
+}
+
+.techTag {
+  background: #f8f9fa;
+  color: $primary-purple;
+  padding: 0.25rem 0.75rem;
+  border-radius: $border-radius-xl;
+  font-size: 0.8rem;
+  font-weight: 500;
+  border: 1px solid #e9ecef;
+}
+
+.projectLinks {
+  display: flex;
+  position: absolute;
+  bottom: $spacing-lg;
+  left: $spacing-lg;
+  gap: $spacing-sm;
+
+  @include tablet {
+    flex-direction: column;
+    gap: 0.75rem;
+  }
+}
+
+.projectLink {
+  padding: $spacing-sm $spacing-md;
+  height: 100%;
+  border-radius: 6px;
+  text-decoration: none;
+  font-size: 0.9rem;
+  font-weight: 600;
+  transition: $transition-normal;
+  border: 2px solid #e9ecef;
+  color: $text-secondary;
+
+  &:hover {
+    transform: translateY(-2px);
+    border-color: $primary-purple;
+    color: $primary-purple;
+  }
+
+  &.primary {
+    background: linear-gradient(135deg, $primary-purple 0%, #764ba2 100%);
+    color: white;
+    border-color: transparent;
+
+    &:hover {
+      box-shadow: 0 5px 15px rgba(102, 126, 234, 0.3);
+      color: white;
+    }
+  }
+
+  @include tablet {
+    text-align: center;
+  }
+}
