+++ apps/home/src/app/pages/project.module.scss
@@ -1,6 +1,8 @@
+@use '../styles/mixins' as *;
+
 .projectTitleContainer {
   display: flex;
-  align-items: baseline;
+  align-items: center;
   gap: 1.5rem;
   margin-bottom: 4rem;
   flex-wrap: wrap;
