+++ apps/home/src/app/styles/_mixins.scss
@@ -1,0 +1,121 @@
+@use 'variables' as *;
+@forward 'variables';
+
+// Responsive breakpoints
+@mixin mobile {
+  @media (max-width: $breakpoint-mobile) {
+    @content;
+  }
+}
+
+@mixin tablet {
+  @media (max-width: $breakpoint-tablet) {
+    @content;
+  }
+}
+
+@mixin desktop {
+  @media (min-width: $breakpoint-desktop) {
+    @content;
+  }
+}
+
+// Flexbox utilities
+@mixin flex-center {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+@mixin flex-between {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+}
+
+@mixin flex-column {
+  display: flex;
+  flex-direction: column;
+}
+
+// Glassmorphism effect
+@mixin glassmorphism($opacity: 0.8) {
+  background: rgba(0, 0, 0, $opacity);
+  backdrop-filter: blur(10px);
+  border: 1px solid $border-light;
+}
+
+// Button styles
+@mixin button-base {
+  border: none;
+  cursor: pointer;
+  transition: $transition-normal;
+  @include flex-center;
+}
+
+@mixin button-round($size: 50px) {
+  @include button-base;
+  width: $size;
+  height: $size;
+  border-radius: $border-radius-full;
+}
+
+// Card styles
+@mixin card-base {
+  background: $background-white;
+  border-radius: $border-radius-lg;
+  box-shadow: $shadow-sm;
+  transition: $transition-normal;
+}
+
+@mixin card-hover {
+  &:hover {
+    transform: translateY(-2px);
+  }
+}
+
+// Text styles
+@mixin text-gradient($color1: $primary-purple, $color2: #764ba2) {
+  background: linear-gradient(135deg, $color1, $color2);
+  -webkit-background-clip: text;
+  -webkit-text-fill-color: transparent;
+  background-clip: text;
+}
+
+// Animations
+@mixin fade-in($duration: $transition-normal) {
+  opacity: 0;
+  transition: opacity $duration;
+
+  &.visible {
+    opacity: 1;
+  }
+}
+
+@mixin slide-up($distance: 20px, $duration: $transition-normal) {
+  transform: translateY($distance);
+  opacity: 0;
+  transition: transform $duration, opacity $duration;
+
+  &.visible {
+    transform: translateY(0);
+    opacity: 1;
+  }
+}
+
+// Hover effects
+@mixin hover-scale($scale: 1.05) {
+  transition: transform $transition-normal;
+
+  &:hover {
+    transform: scale($scale);
+  }
+}
+
+@mixin hover-lift {
+  transition: transform $transition-normal, box-shadow $transition-normal;
+
+  &:hover {
+    transform: translateY(-4px);
+  }
+}
