+++ apps/home/src/app/styles/_variables.scss
@@ -1,0 +1,51 @@
+// Colors
+$primary-purple: #667eea;
+$primary-purple-light: rgba(102, 126, 234, 0.1);
+$primary-purple-border: rgba(102, 126, 234, 0.2);
+
+$text-primary: #333;
+$text-secondary: #666;
+$text-light: #999;
+
+$background-white: #fff;
+$background-dark: rgba(0, 0, 0, 0.8);
+$background-overlay: rgba(0, 0, 0, 0.3);
+
+$border-light: rgba(255, 255, 255, 0.1);
+$border-medium: rgba(255, 255, 255, 0.2);
+
+// Spacing
+$spacing-xs: 0.25rem;
+$spacing-sm: 0.5rem;
+$spacing-md: 1rem;
+$spacing-lg: 1.5rem;
+$spacing-xl: 2rem;
+$spacing-xxl: 3rem;
+
+// Border radius
+$border-radius-sm: 8px;
+$border-radius-md: 12px;
+$border-radius-lg: 16px;
+$border-radius-xl: 20px;
+$border-radius-full: 50%;
+
+// Shadows
+$shadow-sm: 0 2px 4px rgba(0, 0, 0, 0.1);
+$shadow-md: 6px 6px 12px rgba(0, 0, 0, 0.3);
+$shadow-lg: 0 20px 40px rgba(0, 0, 0, 0.3);
+
+// Transitions
+$transition-fast: 0.2s ease;
+$transition-normal: 0.3s ease;
+$transition-slow: 0.5s ease;
+
+// Breakpoints
+$breakpoint-mobile: 480px;
+$breakpoint-tablet: 768px;
+$breakpoint-desktop: 1024px;
+
+// Z-index
+$z-index-base: 1;
+$z-index-dropdown: 10;
+$z-index-sticky: 100;
+$z-index-modal: 1000;
